Add initial project
The initial project contains creates cacti that can fall and collide with a each other as well as a solid platform.

Use the 'Q', 'W', 'E', 'R', and 'T' to make each of the cacti jump, from bottom to top, respectively.

When cactus (A) lands on top of cactus (B), cactus (B) becomes the parent of cactus (A), thus allowing cactus (A)'s transforms to perform relative to cactus (B)'s transforms.